pyproject.toml
[project]
name = "gpt-powered-application-tracker"
version = "0.1.0"
description = "GPT-Powered Application Tracker is a Python application that uses OpenAI's GPT model to track job applications, providing a user-friendly interface and integration with Google Sheets for data storage."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "flet[all]==0.28.3",
    "google-auth>=2.40.2",
    "gspread>=6.2.1",
    "openai>=1.82.1",
    "playwright>=1.52.0",
]
